Redo rendering
Setting pixels directly doesn't work if the surface is not locked, and
filling rectangles doesn't work if it is locked. Additionally, there are
issues with writing past the pixel buffer and messing up other memory,
so it's best to manually write to the pixel buffer in all cases and
check for buffer overflows early and often.

This fixes many of the bugs I was encountering, but the texture still
seems to get messed up. Still the crashes seem to be mostly, if not
completely, gone.